AUTOMATIZACAO/autoNoCode LIXO/test_patch_typings.py
import logging
import os
import subprocess
import sys
from typing import List
import git  # Certifique-se de ter o módulo gitpython instalado

logging.basicConfig(level=logging.INFO)

class AutoTyper:
    def __init__(self, repo_path: str) -> None:
        self.repo_path = repo_path

    def run(self) -> None:
        try:
            self.create_backup()
            self.create_git_branch()
            self.add_type_annotations()
        except Exception as e:
            logging.error(f"Erro durante o processo: {e}")
            sys.exit(1)

    def create_backup(self) -> None:
        logging.info("Criando backup do diretório src...")
        backup_command = 'cp -r src src_backup'
        result = subprocess.run(backup_command, shell=True)
        if result.returncode != 0:
            raise subprocess.CalledProcessError(result.returncode, backup_command)

    def create_git_branch(self) -> git.Repo:
        logging.info("Criando novo branch no repositório Git...")
        branch_name = 'add-type-annotations'
        repo = git.Repo(self.repo_path)
        if branch_name in repo.branches:
            logging.warning(f"A branch '{branch_name}' já existe. Continuando...")
        else:
            repo.git.checkout('-b', branch_name)
            logging.info(f"Criado branch: {branch_name}")
        return repo

    def add_type_annotations(self) -> None:
        logging.info("Adicionando anotações de tipo aos arquivos Python...")
        # Adicione a lógica para adicionar anotações de tipo aqui

def main() -> None:
    if len(sys.argv) < 2:
        logging.error("Por favor, forneça o caminho para o repositório.")
        sys.exit(1)

    repo_path = sys.argv[1]
    auto_typer = AutoTyper(repo_path)

    dry_run = '--dry-run' in sys.argv
    if dry_run:
        logging.info("Executando em modo de teste (dry run). Nenhuma alteração será feita.")
    else:
        auto_typer.run()

if __name__ == "__main__":
    main()
