pyproject.toml
[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"

[tool.poetry]
name = "autonocode"
version = "0.1.0"
description = "AutoNoCode Project"
authors = ["LLPCF <95063971+LLPCF@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typing-extensions = "^4.0.0"
flask = "*"
fastapi = "*"
uvicorn = "*"
sqlalchemy = "*"
psycopg2 = "*"
requests = "*"
fastapi-users = "*"
fastapi_users_db_beanie = "3.0.0"
fastapi-users-db-sqlalchemy = "*"
a2wsgi = "1.10.6"
aiohttp = "3.9.5"
aioquic = "1.1.0"
astropy = "6.1.1"
av = "12.2.0"
brotli = "1.1.0"
brotlicffi = "1.1.0.0"
ca_certs_locater = "1.0"
checks = "0.2"
ConfigParser = "7.0.0"
contextlib2 = "21.6.0"
curio = "1.6"
cx_Oracle = "8.3.0"
Cython = "3.0.10"
defusedxml = "0.7.1"
disco = "1.40.4"
dl = "0.1.0"
docutils = "0.21.2"
eval_type_backport = "0.2.0"
exceptiongroup = "1.2.1"
filelock = "3.15.4"
funcsigs = "1.0.2"
gitdb_speedups = "0.1.0"
google_auth_oauthlib = "1.2.0"
grpcio-status = "1.39.0"
moviepy = "*"
pyright = "*"

[tool.poetry.dev-dependencies]
pytest = "^8.2.2"
mypy = "^1.10.1"
autopep8 = "^2.3.1"
isort = "^5.13.2"
pyannotate = "^1.2.0"
gitpython = "^3.1.43"
pylint = "^2.17.5"
pyright = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
norecursedirs = "node_modules test_output.txt"
testpaths = ["tests"]

[tool.pyright]
include = ["src", "tests"]
exclude = ["**/node_modules", "**/__pycache__", "src_backup"]
strict = ["src", "tests"]
typeCheckingMode = "strict"
reportMissingImports = true
reportGeneralTypeIssues = true
